Add columns and cards ordering

+++ src/cards/cards.service.ts
@@ -1,4 +1,8 @@
-﻿import { Injectable, NotFoundException } from '@nestjs/common';
+﻿import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Card } from './card.entity';
 import { Repository } from 'typeorm';
@@ -35,6 +39,7 @@ export class CardsService {
     card.title = createCardDto.title;
     card.description = createCardDto.description;
     card.column = await this.boardColumnsService.findOneById(columnId);
+    card.position = (await this.findAllByColumnId(columnId)).length + 1;
 
     return await this.cardsRepository.save(card);
   }
@@ -52,13 +57,90 @@ export class CardsService {
     if (updateCardDto.title) card.title = updateCardDto.title;
     if (updateCardDto.description) card.description = updateCardDto.description;
     if (updateCardDto.columnId) card.column = column;
+    if (updateCardDto.position) {
+      const newPosition = updateCardDto.position;
+      const oldPosition = card.position;
+
+      card.position = newPosition;
+
+      await this.updateCardsOrder(
+        card.columnId,
+        newPosition,
+        oldPosition,
+        card.id,
+      );
+    }
+
     return await this.cardsRepository.save(card);
   }
 
+  private async updateCardsOrder(
+    columnId: number,
+    newPosition: number,
+    oldPosition: number,
+    excludeCardId: number,
+  ): Promise<void> {
+    const cards = await this.findAllByColumnId(columnId);
+
+    if (newPosition < 0 || newPosition > cards.length)
+      throw new BadRequestException(`Invalid card position: ${newPosition}`);
+
+    if (newPosition === 0) {
+      newPosition = cards.length + 1;
+    }
+
+    if (newPosition > oldPosition) {
+      this.moveCardsUp(newPosition, oldPosition, cards, excludeCardId);
+    } else if (newPosition < oldPosition) {
+      this.moveCardsDown(newPosition, oldPosition, cards, excludeCardId);
+    }
+
+    await this.cardsRepository.save(cards);
+  }
+
+  private moveCardsUp(
+    newPosition: number,
+    oldPosition: number,
+    cards: Card[],
+    excludeCardId: number,
+  ): void {
+    let card: Card;
+    for (card of cards) {
+      if (card.id === excludeCardId) continue;
+
+      if (card.position > oldPosition && card.position <= newPosition) {
+        card.position--;
+      }
+    }
+  }
+
+  private moveCardsDown(
+    newPosition: number,
+    oldPosition: number,
+    cards: Card[],
+    excludeCardId: number,
+  ): void {
+    let card: Card;
+    for (card of cards) {
+      if (card.id === excludeCardId) continue;
+
+      if (card.position < oldPosition && card.position >= newPosition) {
+        card.position++;
+      }
+    }
+  }
+
   async delete(id: number): Promise<void> {
+    const card = await this.cardsRepository.findOneBy({ id });
+    if (card === null) {
+      throw new NotFoundException(`Column with id ${id} not found`);
+    }
+
     const result = await this.cardsRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Card with id ${id} not found`);
     }
+
+    await this.updateCardsOrder(card.columnId, 0, card.position, card.id);
   }
 }
